requirements.txt
blinker==1.7.0
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.4
MarkupSafe==2.1.3
openai==1.3.5
python-dotenv==1.0.0
redis==4.1.1
requests==2.31.0
urllib3==2.1.0
Werkzeug==3.0.1
